Replace `Button` component (#447)
