+++ packages/web-ui/src/Button/Button.props.ts
@@ -1,5 +1,5 @@
-import { BaseButtonProps } from '../../BaseButton';
-import { Responsive } from '../../types';
+import { BaseButtonProps } from '../BaseButton';
+import { Responsive } from '../types';
 
 export type ButtonProps = BaseButtonProps & {
   /**
