+++ packages/web-ui/src/index.ts
@@ -6,7 +6,6 @@ export * from './types';
 export * from './hooks';
 export { mediaQueries, px, pxToRem, spacing, withGlobalPrefix, DATA_ATTRIBUTES } from './utils';
 /* components */
-export * as Lab from './lab';
 export * from './Alert';
 export * from './Badge';
 export * from './Box';
