--- packages/web-ui/src/lab/Button/index.ts
@@ -1,2 +1,0 @@
-export { Button } from './Button';
-export type { ButtonProps } from './Button.props';
